% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_channels.R
\name{n_channels}
\alias{n_channels}
\title{Recommends number of sampling channels}
\usage{
n_channels(spatial_df)
}
\arguments{
\item{spatial_df}{SpatialPolygonsDataFrame that includes your observational unit. This object should contain all observational units in your geography (e.g. counties in the US), not just those with complete cases.}
}
\description{
Recommends number of sampling channels to use when taking spatial first differences. Please see: Druckenmiller & Hsiang (2018).
}
\examples{
n_channels()
}
\keyword{SFD}
