% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDD_vars.R
\name{SDD_vars}
\alias{SDD_vars}
\title{Generate SFD and SDD variables}
\usage{
SDD_vars(spatial_df, n_channel, obs_var, dependent_var, independent_vars,
  rotation = 0, needs_balanced = FALSE, balanced_df = NA, plot = TRUE)
}
\arguments{
\item{spatial_df}{SpatialPolygonsDataFrame that includes your observational unit, dependent variable, and independent variable(s). This object should contain all observational units in your geography (e.g. counties in the US), not just those with complete cases.}

\item{n_channel}{numeric that specifies the number sampling channels you would like to use when calculating SFD. This number should be chosen such that the height of your sampling channels is approximately equal to the height of your observational units}

\item{obs_var}{obseravational unit}

\item{dependent_var}{dependent variable}

\item{independent_vars}{independent variable(s)}

\item{rotation}{numeric that specifies the angle at which SFD will be taken (-90 to 90 degrees), default=0.}

\item{needs_balanced}{logical; if TRUE, the regression will be conducted on balanced data}

\item{balanced_df}{OPTIONAL; if needs_balancced==TRUE, specify a data.frame containing only those observations included in the balanced data. This data frame must include your observation variable.}

\item{plot}{logical; if TRUE, it will plot the adjacent observations used to calculate SFD in your sample}
}
\description{
Transforms variables by taking spatial first difference (SFD) and spatial double differences (SDD). Please see: Druckenmiller & Hsiang (2018).
}
\examples{
SDD_vars()
}
\keyword{SFD}
