% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfd-package.R
\docType{package}
\name{sfd-package}
\alias{sfd}
\alias{sfd-package}
\title{sfd: Space-Filling Design Library}
\description{
A collection of pre-optimized space-filling designs, for up to ten parameters, is contained here. Functions are provided to access designs described by Husslage et al (2011) \doi{10.1007/s11081-010-9129-8} and Wang and Fang (2005) \doi{10.1142/9789812701190_0040}. The design types included are Audze-Eglais, MaxiMin, and uniform.
}
\author{
\strong{Maintainer}: Max Kuhn \email{mxkuhn@gmail.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

}
\keyword{internal}
